feat: Enhanced Django admin interface for food app
In this commit, the Django admin interface for the food app has been significantly improved. The Ingredient and Item models were added as inline forms under the Recipe model to enhance readability and ease of data manipulation.

The `__str__` methods were added to the Recipe, Ingredient, and Item models for better readability in the admin drop-down menus.

In the Makefile, a command was added to create dummy data for testing and development. The code was cleaned up by removing unnecessary comments in tests.py.