refactor: Refactor admin and test models in food app
This commit refactors the admin and test models in the food application. It ensures that the models are registered using the '@admin.register()' decorator for more inline registration. Additionally, it has formated the exclusion fields in 'IngredientAdmin' and 'RecipeAdmin' classes for simpler updating in the future. Finally, on the tests.py file, a new line was added for line-spacing consistency.